Cleanup PDF sample player
