+++ Palace/Samples/AudiobookSamplePlayer/AudiobookSample.swift
@@ -1,0 +1,14 @@
+//
+//  AudiobookSample.swift
+//  Palace
+//
+//  Created by Maurice Carrier on 9/14/22.
+//  Copyright © 2022 The Palace Project. All rights reserved.
+//
+
+import Foundation
+
+struct AudiobookSample: Sample {
+  var url: URL
+  var type: SampleType
+}
