--- Palace/Samples/EpubSamplePlayer.swift
@@ -1,21 +1,0 @@
-//
-//  EpubSamplePlayer.swift
-//  Palace
-//
-//  Created by Maurice Carrier on 8/14/22.
-//  Copyright © 2022 The Palace Project. All rights reserved.
-//
-
-import SwiftUI
-
-struct EpubSamplePlayer: View {
-  var sample: EpubSample
-
-  var body: some View {
-      contentView
-  }
-
-  @ViewBuilder var contentView: some View {
-    Text("Show Epub reader")
-  }
-}
